traffic_ai_detection/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
opencv-python==4.5.3.56
numpy==1.21.2
ultralytics==8.0.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0